==> analysis/00_data.qmd <==
---
title: "Data provenance"
#author: "Luisa M. Mimmi"
#date: "Last run: `r format(Sys.time(), '%F')`"
lang: en
editor: source
engine: knitr
tbl-cap-location: top
format:
  html:
    # theme: flatly #spacelab
    code-fold: false # redundant bc echo false 
    toc-depth: 2
    toc_float: true
    toc-location: left
    toc-title: Outline
    embed-resources: true # external dependencies embedded (Not in ..._files/)
  # pdf:
  #   toc-depth: 2
  #   toc-title: Indice
  #   highlight-style: github
  #   #lang: it
  #   embed-resources: true # external dependencies embedded (Not in ..._files/)
format-links: false
---

## Source of WB operations' data

The **raw** and **derived** data files are accessible from the [Github Repo {{< fa brands github >}}](https://github.com/Lulliter/slogan) for this project. 

## WB Projects & Operations
**World Bank Projects & Operations** were obtained from: 

+ [World Bank Projects & Operations Data Catalog](https://datacatalog.worldbank.org/search/dataset/0037800) - links + some metadata
+ [World Bank Projects & Operations Advanced Search](https://projects.worldbank.org/en/projects-operations/project-search) - search engine with filters.  

The Accessibility Classification is **public** under [Creative Commons Attribution 4.0](https://datacatalog.worldbank.org/public-licenses?fragment=cc)
 
### Process to ingest & preprocess raw PDO text data

1. Retrieve ALL WB projects (22,571) listed (approval obtained or requested *between FY 1947 and 2026* as of *31/08/2024*) using the `Excel button` on this page: [WBG Projects](https://projects.worldbank.org/en/projects-operations/projects-list?str_fiscalyear=1979&end_fiscalyear=1979&os=0) 
2. Split the dataset and keep only `projs_train` (~50% of projects with PDO text, i.e. **5,637 PDOs**)
3. Clean the `projs_train` dataset 
4. Further processing of the column `pdo`...

<!-- ::: {.callout-warning icon="false" collapse="true" style="background-color: #fffcf9;"} -->
<!-- ### {{< bi terminal-fill color=#9b6723 >}} Nerdy Note -->
<!-- I had made a prior attempt to ingest the WB document data using the `rvest` package, following the work of [Stephanie Tran](https://github.com/transteph/world-bank-document-scraping) project who created the function `R/f_scrape_WB-OKR.R`. However the WB website has changed and the function is not working anymore. -->
<!-- ::: -->

```{r}
#| echo: false
#| warning: false
#| message: false

library(knitr)
library(kableExtra)
```

## WB World Development Reports (WDRs)

**World Development Reports** can be retrieved as a *sub-collection* of the `Corporate Flagship` *Collection* of World Bank's documents. 

At this link:

<!-- `wdr.rds` was created in old repo `/Github/slogan_old/` in `OLD/_my_stuff/WDR-data-ingestion.Rmd` with text processing on WDR abstracts in `OLD/01b_WDR_data-exploration_abstracts.Rmd` and saved as `...OLD/data/raw_data/wdr.rds`. -->

+ [World Development Report](https://openknowledge.worldbank.org/collections/3d9bbbf6-c007-5043-b655-04d8a1cfbfb2) - there are 44 items available as of today (`r format(Sys.time(), '%F')`).

I initially managed to use an API call to retrieve the WDRs metadata and save it (in 2022) following [Stephanie Tran's](https://github.com/transteph/world-bank-document-scraping) project who created the function `R/f_scrape_WB-OKR.R`. However, due to changes in url, I could not repeat that process. 

See more info here:

+ [World Bank APIs](https://datahelpdesk.worldbank.org/knowledgebase/articles/889386-developer-information-overview) 

+ [Third party APIs](https://data.worldbank.org/products/third-party-apps) 


***

## 0) Input data files 
These files in the folder `data/raw_data/` are downloaded from the World Bank website.


```{r}
#| label: tbl-input_files
#| echo: false
#| warning: false
#| message: false
#| asis-output: true

data.frame(
  `Source File Name` = c(
    "project2/all_projects_as_of29ago2024.xls",
    "project3/all_projects_as_of31mar2025.xlsx (Sheet Projects)",
    "project3/all_projects_as_of31mar2025.xlsx (Sheet Themes)",
    "project3/all_projects_as_of31mar2025.xlsx (Sheet Sectors)",
    "project3/all_projects_as_of31mar2025.xlsx (Sheet GEOLocations)",
    "project3/all_projects_as_of31mar2025.xlsx (Sheet Financers)",
    "wdr.rds"
  ) |>
    (\(x) sprintf('<code style="font-family: monospace; font-size: 85%%">%s</code>', x))(),

  Details = c(
    "22,571 obs (projects)",
    rep("22,210 obs (projects)", 5),
    "45 obs (WDRs)"
  ),

  Retrieved = c(
    "29 of August 2024",
    rep("31 of March 2025", 5),
    "30 of June 2022 + manual additions in October 2024"
  ),
  check.names = FALSE
) |>
  kable(
    format = "html",
    escape = FALSE,
    caption = "List of source files and retrieval dates"
  ) |>
  kable_styling(
    full_width = FALSE,
    bootstrap_options = c("striped", "hover")
  ) |>
  column_spec(1, width = "25em")

```


## 1) Output data files from text preprocessing and EDA
These files in the folder `data/derived_data/` are created in different scripts and saved here to be reused in other scripts.


```{r}
#| label: tbl-ouput_files01
#| echo: false
#| message: false
#| warning: false
#| asis-output: true

data.frame(
  `File <code>*.rds</code> name` = c(
    "<code>wdr.rds</code>",
    "<code>all_proj_t.rds</code>",
    "<code>projs_train.rds</code>",
    "<code>projs_test.rds</code>",
    "<code>projs_val.rds</code>",
    "<code>pdo_train_to_tag.rds</code>",
    "<code>pdo_train_tagged.rds</code>",
    "<code>pdo_train_t.rds</code>",
    "<code>projs_train2.rds</code>",
    "<code>pdo_train2_t.rds</code>",
    "<code>custom_stop_words.rds</code>",
    "<code>custom_stop_words_df.rds</code>",
    "<code>wdr2.rds</code>"
  ),
  
  `Source File Name` = c(
     "[from <code>OLD</code> repo <code>~/Github/slogan_old/</code>]",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01a_WB_project_pdo_prep.qmd</code>",
     "<code>analysis/01b_WB_project_pdo_EDA.qmd</code>",
     "<code>analysis/01b_WB_project_pdo_EDA.qmd</code>",
     "<code>analysis/01b_WB_project_pdo_EDA.qmd</code>",
     "<code>analysis/01b_WB_project_pdo_EDA.qmd</code>",
     "<code>analysis/01b_WB_project_pdo_EDA.qmd</code>"
  ),
  
  Details = c(
     paste("as df (44)", "<br>", "<em>problem with changed API</em>"),
     "11,279 obs (projects)",
     "5,637 obs (projects)<br>4,425 if &lt; 2001 FY",
     "2,821 obs (projects)",
     "2,820 obs (projects)",
     "5,637 obs (input)<br>Post split",
     "LARGE `cnlp` object<br> <em>intermediate output</em>",
     "314,821 obs (tokens)<br>(248,256 if &lt; 2001 FY)",
     "4,425 obs (projects)<br><em>changed</em>",
     "252,705 obs (tokens)<br><em>changed</em>",
     "as vector",
     "as df",
     "as df (46)<br>[added WDR 2023/2024 manually]"
  ),
  
  check.names = FALSE
) |>
  kable(
    format = "html",
    escape = FALSE,
    caption = "Clean `.rds` files with their source and contents"
  ) |>
  kable_styling(
    full_width = FALSE,
    bootstrap_options = c("striped", "hover")
  ) |>
  column_spec(1, width = "14em") |>
  column_spec(2, width = "30em") |>
  column_spec(3, width = "25em")  

```



## 2) Output data files from ML classification

```{r}
#| label: tbl-output_files02
#| echo: false
#| message: false
#| warning: false
#| asis-output: true

data.frame(
  `File <code>*.rds</code> name` = c(
    "<code>.....rds</code>" 
  ),
  
  `Source File Name` = c(
     "<code>analysis/02a_WB_project_pdo_feat_class_envcat.qmd</code>" 
  ),
  
  Details = c(
     paste("..... ", "<br>", "<em>xxxx </em>") 
  ),
  
  check.names = FALSE
) |>
  kable(
    format = "html",
    escape = FALSE,
    caption = "Clean `.rds` files with their source and contents"
  ) |>
  kable_styling(
    full_width = FALSE,
    bootstrap_options = c("striped", "hover")
  ) |>
  column_spec(1, width = "14em") |>
  column_spec(2, width = "30em") |>
  column_spec(3, width = "25em")  

```
